+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 * `hasSettings` -> `hasCpSettings` for Craft 3 beta 8 compatibility
 * Added Craft 3 beta 8 compatible settings
+* Modified config service calls for Craft 3 beta 8
 
 ## 1.2.3 - 2017.03.12
 ### Added